Rename `insert_anywhere` to `record_bind_in_anywhere`
Summary:
The name `insert_anywhere` was confusing because it lived alongside other
functions where "insert_xyz" meant "add an xyz to bindings", whereas in this
case it meant "insert a w if necessary, then add a thing to the w" which is
a more complex operation.

Giving it a more descriptive name makes it easier to see at a glance how
the higher-level `bind_name` function actually works, because now it's easier
to see that we're passing the binding we just created to be recorded in an
Anywhere binding.

Also added a doc comment, since the operation is complicated enough not to
be self-explanatory to anyone who doesn't already know about our notion
of Anywhere.

Reviewed By: grievejia

Differential Revision: D82548821

fbshipit-source-id: ee353619fa298df07b36bd3bd3b36aeb72d6fea8