Use a type to represent read lookups
Summary:
## About this stack

See D82378108 for full details - I'm trying to unblock a change to flow-merging by first
re-structuring the scope so that there is just one map of information about names currently
tracked by the scope, and the elements of that map can distinguish between relevant cases:
- Uninitialized locals
- Initialized (or possibly-initialized) locals, including both assigned type and any narrows
- Mutable captures, including any narrows

To make this happen without a single massive change to the entire bindings
layer, I first need to encapsulate the details of the current `FlowInfo` / `StaticInfo`
layout behind higher-level APIs that more clearly describe the kinds of things other
parts of bindings are actually trying to do.

## About this sub-stack

In this sub-stack, I make the scope stack own how to look up the currently bound key
for a name, by cleanly separating that operation (which is a pure function and depends only
on scope semantics) from the mutations involved in intercepting first usage (which is a binding
table concern, not something modeled in scope) and mutating the tables (again, a binding table
concern).

## About this diff

As promised in a TODO, this diff makes it clearer what the name read cases are by using
a documented enum type instead of `Result<Either<...>, ...>`. There is no functional change.

Reviewed By: grievejia

Differential Revision: D82498636

fbshipit-source-id: 2c4a3cfa8fb32fd96df07afc801c9d1e907b65c5