Document that fresh quantified vars need to be finalized
Summary:
To address feedback on D82493518, this diff tries to make it more clear that quantified vars need to be finalized.

* Change finish_quantified to consume the vars, instead of taking a reference.
* Wrap the vars in a handle annotated with `must_use`

The `must_use` attribute doesn't work as well as I'd like, because we return the handle as part of a tuple. For now, it's at least nice documentation.

In the future, it might be nice to change fresh_quantified to return a single value that wraps the type and quantified vars, to make it harder to forget to finalize the vars by making it harder to get at the type.

Reviewed By: stroxler

Differential Revision: D82553475

fbshipit-source-id: 77a52f448b9afe1661617689e0f657a8a2753771