Add django stubs to test env (#1101)
Summary:
Pull Request resolved: https://github.com/facebook/pyrefly/pull/1101

Added stubs for models in a third-party dir, as well as the license file.

This directory contains many useful stubs which I noticed in the testsuite, including field definitions, enums, as well as the Model definition.

There were some subdirectories in the original stubs which I removed until I know for sure if we need them.

Reviewed By: samwgoldman

Differential Revision: D82494385

fbshipit-source-id: 269507ad54a36801a4ca604aaf63ca3e750eda28