Return `TypeErrorDisplayStatus::DisabledDueToMissingConfigFile` when the file is covered by a marker config
Summary:
When I was initially writing this logic, I forget to account for the mypy and friends case: we don't parse these configs, but we just use them to infer a project root but otherwise use a default config. These configs will have type checking off, and have `Marker` as the config source. They will lead to the confusing "Pyrefly type checking is enabled through a config file." message when the user hover on "Pyrefly (error-off)".

This diff fixes the case. We will report `TypeErrorDisplayStatus::DisabledDueToMissingConfigFile` in this case, same to the case where we have no configs.

Reviewed By: kinto0

Differential Revision: D82575438

fbshipit-source-id: 8e3b00cc58f6e11859a492bc43397cf616bf55fc