Explain mutable capture logic better
Summary:
The logic where we both mutate and add a capture to static is pretty
confusing, but it's also very hard to change in the current structure
and we need to refactor to replace all this code anyway.

In the meantime, it might be helpful to explain it in a comment. While
I'm at it I shorten the name of `lookup_mutable_capture`, which seemed
unnecessarily verbose.

Reviewed By: samwgoldman

Differential Revision: D82585432

fbshipit-source-id: 7b77eb09de6e39705eb2c281350fec2b323deb1f