Simplify removing a name from scope
Summary:
An oddity of the trunk flow behavior in Pyrefly is that we have no way
to remove a name from the current flow cleanly.

In a way this makes sense: we can't use an actual remove becuase if the name
*were* defined we would want the type to still be visible for best-effort
typing downstream, so the operation we want is to set the style to
`Uninitialized`.

But at the moment the only way to do that is an upsert, which leads to weird
behavior where we'll actually *insert* a nonexistent name into the current flow
on a `del` of an unknown name.

If we just make a direct helper that does what we want - set the style
to `Uninitialized` if the name exists and ignore it otherwise - a few
operations become simpler and less code is coupled to details of how
flow style is represented.

Reviewed By: samwgoldman

Differential Revision: D82585434

fbshipit-source-id: aa1e549807e277dc37511988e5f9caf1f843430c