Track mutable capture in `definitions.rs`
Summary:
We're finally ready to take our first big step in the direction of
handling mutable captures more clearly: we ought to actually track
in `definitions.rs` when a name is a mutable capture.

By doing so, we're immediately able to simplify several things:
- We no longer need to do a bind `add_to_static` call in
  the handler for "declaring a mutable capture", because we no
  longer have to worry that the static environment.
- We no longer need an inconsistently-tracked `is_mutable_capture`
  flag in `StaticInfo`, because now we can get this directly from
  the definition style.
- The `add_to_static` interface is now simpler since we aren't
  passing the now-unneeded flag.

In addition, this opens the door to refactoring `StaticInfo` in
a way that would let us more clearly model what's going on and
fix some of the open bugs.

Note that the ordering of cases in the definition style is significant - we
have defined a `global` to take precedence over a `nonlocal` and either of then
to take precedence over all other definition styles. That's because regardless
of what operation would bind a name, if it has been marked as a mutable capture
then that name affects an enclosing scope, not the current scope (and if this
is a top-level scope, it's always going to be both a type error and a
compile error at runtime, so we can largely ignore that case).

Reviewed By: samwgoldman

Differential Revision: D82585433

fbshipit-source-id: dcd03e6757b83c2f7215191b37dbf21fb5634515