Validate typevar instantiations on calls
Summary:
Adds enough structure to be able to emit errors including the expected and actual types and typevar name. I tried to re-use the same code for explicit specialization as much as possible.

This improves conformance, but adds 1 additional difference. We promote literals when instantiating vars, which is wrong when the typevar bound is a literal. This is already documented in a test, test_typevar_literal_bound.

Fixes https://github.com/facebook/pyrefly/issues/111

Reviewed By: grievejia

Differential Revision: D82589554

fbshipit-source-id: 792dec326f4c2880d0405350ffab28d894e0c5c2