Don't promote when instantiating typevar if the typevar bound is literal
Summary:
This diff takes a guess-and-check approach, which I think is fine because errors should be rare enough that a less-than-optimized approach is fine.

Essentially, we try to promote. If that's incompatible with the bound, we try again with the un-promoted type. If both fail, we error using the promoted type.

Reviewed By: grievejia

Differential Revision: D82590745

fbshipit-source-id: a13b836089a8be9446624dee65755b6e2b5f67db