Don't promote literals inside callables
Summary:
`abstractmethod` is typed as an identity function, and we promote literal types passed to an identity function, so decorating any function with `abstractmethod` was causing literal types in the signature to be promoted.

I can't think of a case in which we would want to promote literals in a callable signature, so I changed promote_literals() to skip literals inside callables. I didn't find an existing variant of transform() that could do this, so I copied the body of transform_mut() and tweaked it to not recurse into callables.

Fixes https://github.com/facebook/pyrefly/issues/1058.

Reviewed By: ndmitchell

Differential Revision: D82608178

fbshipit-source-id: ab07e857d054a7a90373aed3dddb73c0c0f62078