Rename `mark_as_not_in_current_flow`, tweak comment
Summary:
Follow-up from discussion on D82585434.

I also decided the comment could be more explicit that Python behaves
strangely here - to the best of my knowledge this the only case where
Python attempts to approximate block scoping, and the resulting behavior
can be pretty strange if the exception handler used an already-defined
name (I did verify it in an interpreter to cross-check the docs).

Reviewed By: samwgoldman

Differential Revision: D82655420

fbshipit-source-id: 3e856aee4a2a5585fa666bd8296d0413ad59cdbb