Clean up the `UsageLink` for delete
Summary:
When implementing first-usage pinning, we had to create
special link bindings whenever a `BindingExpect` (or other
non-`Binding` keyed-binding-type) made use of `ensure_expr`,
so that we could track the lookups.

Now that we've moved `Delete` into a `Binding`, we can get
rid of this - the actual delete is directly usable as a
first-usage, no need for a `UsageLink` here.

Note that since `Binding` is bigger than `Type`, this means
converting it to a `Binding` - which might seem wasteful
in the case where the name is defined in some other way,
because we are wasting a `Binding` and `TypeInfo` on the
result - is probably a net win in terms of Ram consumption,
the link wasted more space than the unused answer does.

Reviewed By: grievejia

Differential Revision: D82692656

fbshipit-source-id: b4600bc518163cd0a67454708f83f741f5de1ca4