We should be looking for walrus in `Delete`
Summary:
The current `definitions.rs` code is skipping `del` statements; I suspect
it's because we were imagining the del target is a heavily restricted
target.

But this is actually not the case - it's true that the target of `del` is
restricted, but if the `del` is of an attribute or subscript, the base
sub-expression is arbitrary.

That means we were skipping over `Named` expressions inside the base. This
diff closes the gap.

Reviewed By: grievejia

Differential Revision: D82692659

fbshipit-source-id: 2a13e54a1b50c7284af6f2c17c615fdd5a128f85