Add `del`-defined names to `Definitions`
Summary:
We now more-or-less properly model the impact of `del` on the scope.

The test case that demonstrated issues still doesn't quite behave
right - the errors are inconsistent and not entirely aligned with
semantics becasue of the way we are using `LookupKind::mutable` to
try to validate mutations using read lookups, but things are better
than before and more of the plumbing is now in place to handle
scope better.

Reviewed By: grievejia

Differential Revision: D82692658

fbshipit-source-id: b9c8982586a361b7670eca64aa06e66d96e4b447