Move delete logic to a Binding
Summary:
I want to fix up a bug in scope: `del` defines a local
variable, and needs to be represented as such.

But if a variable is defined only by a `del` (which is
always both a type error and guranteed to crash unless the
code is doing highly dynamic things), we have to provide
well-defined behavior on attempts to read that variable.

A sensible way to do this is to turn the `Delete` check
into a binding, to which we will link the `StaticInfo`
in the case where a variable is defined only by `del`
once we fix the scope logic. That's what this diff does;
we already will produce a type error for the broken `del`
so we use `any_implicit` as the type in order to avoid
redundant errors on reads.

This diff has no behavior change because the binding is
currently unused; it won't affect behavior until we change
the scope logic.

Reviewed By: grievejia

Differential Revision: D82692661

fbshipit-source-id: 5ac7b5650a7fc2a3121e3af5ab656d5c8edaee8d