Factor `check_del_statement` logic out
Summary:
Because of scoping issues that create edge cases where we need to provide a
read type for a local that exists only due to `del` (this is very similar to
what already happens for mutable captures), I want to move the checking of
`Del` to a binding that returns `Any`.

In the process, I think it's worth factoring out the logic (most of which
will remain exactly as-is) so that the big match statements shrink
a bit.

Reviewed By: grievejia

Differential Revision: D82692662

fbshipit-source-id: 3bb962d2b576aab8e4f502dffe67ade496d612b3