Track weird behavior of augmented assign lookups
Summary:
For a long time there's been a hook in the name lookup logic where
mutable lookups and barrier interact. I'm still not sure why that hook
is there, but this test confirms that it has strange side-effects: the
error if we have an augmented assign `x += ...` inside a function with
no other assign to `x` is misleading - we get the error that one would
normally get when `x` isn't defined in any scope.

But `x` actually *is* defined - you can see this because if we `reveal_type`
*above* the augmented assign we instead get the uninitialized-local error
one would normally expect.

The root cause is that `LookupKind::Mutable` is a conceptual error - the
lookup function it is in is a *read* lookup, which shouldn't care about
mutation. Augmented assignments are a read plus a write, and writability
ought to be tracked separately (we could combind the two in a helper
function, but they are still separate issues).

In order to actually make that change we'll need to improve handling of both
augassign and del, both of which are relying on `LookupKind::Mutable` to
catch illegal operations on normal captures in a way that is approximately
right but gets edge cases wrong.

Reviewed By: grievejia

Differential Revision: D82692660

fbshipit-source-id: c111dea395866a8ff4a10711900677199aba11bf