Track the scoping behavior of `del` as a bug
Summary:
Pyrefly arguably mostly supports `del` right now, in the sense that we
do throw errors on code that uses `del` in ways that don't make
sense.

But the errors aren't really right, and expose issues in how we are
representing scope: at runtime, the Python compiler includes `del`
in the forms that define a local variable (unless there's a mutable
capture declaration for the same name).

Pyrefly, on the other hand, treats `del` as not defining a name which means we
produce "could not find" errors insted of uninitialized errors when there's no
shadowing, and we produce "is not mutable" errors when there is shadowing. In
both cases, the errors aren't quite right.

Reviewed By: grievejia

Differential Revision: D82692655

fbshipit-source-id: 501078f4ce778e3769b48996fd3c34aeb5ff82b4