Introducing a `NestingContext` data structure to represent nestings properly
Summary:
We currently don't track the context when a name is nested inside another class/function. For example, for this code
```
# foo/bar.py
class Baz:
  class Qux:
    def test(self): ...
```
We'll print the fully-qualified name of the class `Qux` as just `foo.bar.Qux`, which is clearly wrong -- a layer of class name `Baz` is completely dropped!

This diff introduces the notion of "nesting context" which is a concept I ported from [Pyre1](https://www.internalfb.com/code/fbsource/fbcode/tools/pyre/source/ast/nestingContext.mli). Given an AST node, it simply represents a "trace" of where the node is nested. When there's no nesting, we have the "toplevel" context; when it's nested in a class, we have the "class" context; when it's nested in a function, we have the "function" context. Class and function context can be themselves nested under another context, meaning that the entire "trace" of the context forms a linked list.

Reviewed By: stroxler

Differential Revision: D82697494

fbshipit-source-id: 3689b5b1d195eef93601f55daba2725b0a1ea6e4