Add nesting context to `Class::has_qname()`
Summary:
Now that `QName` has an additional nesting context, we need to change the `has_qname()` API to take that into account.

I renamed the pre-existing `has_qname()` API to `has_toplevel_qname()` since all pre-existing callers assume the name must be defined in the toplevel. The new `has_qname()` now takes in additional `NestingContext` argument so if we want we should be able to perform nesting-context-aware matches.

Reviewed By: stroxler

Differential Revision: D82697734

fbshipit-source-id: 8b87a15632e9dd04eefcf0e3c3c88d53eecaf7ca