Re-structure name lookup logic
Summary:
This is a fairly big diff; I couldn't find a way to make it smaller
easily because I need to both add new plumbing and rip out some
incorrect plumbing, and trying to do it in small bits would just
be more confusing.

## What is the state of the world on trunk

Uninitialized variables can be detected in three different ways:
- We might fail a lookup using `LookupKind::Mutable` and produce a "name
  `whatever` is not mutable from the current scope". This error is a Pyrefly
  phabrication - it is impossible to encounter anything like this from scope
  issues in Python because any operation that mutates a reference also defines
  a local.
- In certain situations, we might instead fail to find a well-defined local at
  all, and instead produce the "name `whatever` not found" error.
- Or, if the lookup succeeds, a separate part of the codebase will also try to
  double-check the current flow style and we might produce uninitialized local
  errors there.

This is pretty confusing - we're using three different errors for something
that is exactly the same semantic issue (and encounters the same
UnboundLocalError at runtime), plus we have two separate lookups and a
LookupKind flag that shouldn't even exist modifying behavior.

## What is the new state of the world

The Scope stack understands the current scope information, and just
hands back an answer with enough detail to handle any logic - including
producing uninitialized local errors - that is needed downstream.

Because the `BindingsBuilder` has to handle several extra complexities
that aren't part of scope per se, we have a separate `NameLookupResult`
that is simpler and represents the final answer after accounting for:
- potentially intercepting legacy tparam lookups
- intercepting first-usage bindings
- taking anywhere-style `Keys` (which may not yet exist, which is why
  `Scopes` has to give a `Key`) and converting them to `Idx<Key>` by
  inserting them into the table.

There's a bit of complexity introduced here where `NameLookupResult`
is generic, which was caused by us aggressively optimizing the
legacy tparams lookups. If we are willing to just use a `Forward`,
then I could get rid of this in a follow-up: the problem is that at
the moment we skip the Forward and use a raw binding, which forces
the lookup result to allow either an `Idx<Key>` or a `Binding`.

At this point, all uninitialized local errors have the same error
code and error message, we no longer are synthesizing a made-up
concept of "not mutable from the current scope". And we only
have one lookup, everything is powered by the information it
provies - we no longer do a separate flow style lookup
in `ensure_name`.

## Context for making the change

This change has benefits in itself:
- It's a big BE win because the data flow makes a lot more sense
- It fixes a genuine false-positives where we could read uninitialized
  locals that shadowed parent scopes without getting an error.
- It fixes multiple test cases I had marked as bugs where we produced
  errors, but with misleading or inconsistent error messages.

All of that would be medium or low priority, the bigger impact is that
by simplifying things it will help unblock bigger scope changes.

My overall medium-term goal here remains to change how scope is represented
and make it easy to track narrows versus assigns separately, so that
our flow merging logic can more aggressively optimize out Phis, which
will fix many bugs and also improve our options for loop handling.

This diff is fixing one of the biggest roadblocks I hit - the name lookup
logic was very complex and also didn't actually make sense, which was
making it very hard for me to visualize what a refactor should look like.

Reviewed By: grievejia

Differential Revision: D82761271

fbshipit-source-id: 78398d2780f48ce86b4319b550e615c8cec546b8