Rename implicit global related things
Summary:
Since `global` is both a well-defined Python concept and is directly relevant
to `Definitions` and to `StaticInfo` (both of which care about mutable
captures) I keep getting a bit mixed up when I see these varible names and
variant names.

If we just label them as implicit, it's much clearer.

Reviewed By: samwgoldman

Differential Revision: D82800315

fbshipit-source-id: a848dac95b42774cdc4bf63fd3e7dba1d69012c5