Add tests showing committable transactions must be executed serially, and there is no deadlock
Summary:
I am trying to write unit tests to repro stuck threads bug. I suspect there might be deadlock, so I wrote this test, but it turns out to be no deadlock, at least with this type of interactions.

However, the test can be helpful anyways, so here is the diff.

Reviewed By: stroxler

Differential Revision: D82753943

fbshipit-source-id: dcb26fe3a57201103edd4f53c4453e2ce5245d8e