Add decorator testcases
Summary: This is also aimed at eliminating false positives. I am noticing that tests are timing out when adding more of them, so I needed to increase the timeout again. It's not one specific test that's failing. It's mostly that again, when we run in parallel, things become slower as we keep adding more tests. I slightly increased the timeout, but would love to know if there are any further thoughts on how to address.

Reviewed By: rchen152

Differential Revision: D82770737

fbshipit-source-id: ef1b447f89bbb929317732afe5d66c560753161a