Add admin failing testcases
Summary:
I am focusing on minimizing false positives, so we at least want to interpret the definitions correctly.

Once we add stubs, one of the testcases will still fail and we should fix it.

For some reason, adding those tests caused a timeout (since tests are slower when run in parallel), so I increased the timeout by one second and it fixed the issue.

Reviewed By: rchen152

Differential Revision: D82766335

fbshipit-source-id: 6fae105d082f6e2020b8048dccdf999df83b3014