Rewrite lookup_legacy_tparam_from_idx for readability
Summary:
This is essentially a rewrite of D82776935. I thought of a more readability way to do the same thing, but I didn't want to touch a diff in the middle of a large stack.

The way I originally modified lookup_legacy_tparam_from_idx to handle both Name and Attr tparams is not very readable, because the "idx" argument ends up meaning different things depending on the value of another argument. Instead, let's factor out a helper function for the common structure of a lookup function and have it take a closure that knows how to do a Name- or Attr-specific lookup.

Reviewed By: stroxler, ndmitchell

Differential Revision: D82868308

fbshipit-source-id: b993a8acc89fd749ff712370b9734c0e0e09cb7c