use information in exports for as_special_form
Summary:
When the user creates a `typing.TypeVar`, we detect it in the bindings step and produce special bindings for it. The list of names we recognize for special handling is defined in SpecialExports. Right now, this only works if `TypeVar` is imported directly from `typing`.

If a module imports `typing.TypeVar` and re-exports it, usages of the re-exported version will not be identified correctly, which will lead to a lot of typechecking errors.

One example of this pattern is in narwhals: https://github.com/narwhals-dev/narwhals/blob/main/narwhals/_typing_compat.py - it's useful to put conditional import logic (if TYPE_CHECKING, typing v.s. typing_extensions based on sys.version) in a custom typing utils file to avoid cluttering every single file.

This diff makes use of the information calculated in the previous diff to support special export lookup for re-exported names.

Fixes https://github.com/facebook/pyrefly/issues/1050

Fixes https://github.com/facebook/pyrefly/issues/925

Reviewed By: yangdanny97

Differential Revision: D82254103

fbshipit-source-id: c40cbeb4dbfb34bb273618a5c31735c9821c80db