Use go-to-definitions as callees
Summary:
In `callGraph.ml` we currently rely on types of callables to get their names.
However this approach is no longer feasible in Pyrefly, since types no longer
contain name information. As a workaround, we can probably begin with
go-to-definitions and see which call graph edges are missing / incorrect, and
then improve from here.

Reviewed By: arthaud

Differential Revision: D82792995

fbshipit-source-id: e17895aca9295f2a72e0bd29bac013fb89cfec69