Use Python AST compatible locations
Summary:
`DisplayPos` and `DisplayRange` exposed by Pyrefly are defining position in term of characters which is different from the way how Python AST sets values for `lineno`, `col_offset`, `end_lineno` and `end_col_offset`. In Python line numbers are 1 based and `col_offset` is 0 based byte offset from the beginning of the string. For ASCII files  offsets match with characters so in a lot of cases we are fine however it breaks for unicode.

This diff switches reported locations to use Python approach and removes sports in py2hack that used to do adjustments

Reviewed By: stroxler

Differential Revision: D82974021

fbshipit-source-id: 685988f1ade8700f7115cc5e4f1eb6f763ae5d31