Simplify `add_parameter_to_current_static`
Summary:
Adhoc add-to-static is tricky to reason about because in general it is
unsafe to use (you might change the count, for example, which would potentially
lead to panics around Anywhere).

We rely on it for parameters though - these are safe to model via
`add_to_current_static` because the scopes they appear in are simple and
statement-free, so the panic conditions don't exist.

So, let's rename the function. While I was at it I realized that it's always used
with an `Identifier`, so we can simplify the interface (which also helps lock down
the semantics - you can't use the function unless you have an actual AST node... this
is a desirable restriction, it would always be wrong to use it without an Identifier).

Reviewed By: samwgoldman

Differential Revision: D82851677

fbshipit-source-id: 77a78ce4a7142f1fdbe998186ff5a93eb76cf403