Rearrange test cases
Summary:
I want to eliminate a bunch of redundant tests - only a relatively small
subset of the existing nonlocal vs global tests are actually useful,
because only the search-enclosing-scopes portion of the logic is actually
different.

So, I'm collecting all the search-enclosing-scopes relevant tests up top,
and I'll delete a bunch of the other redundant ones in the next diff.

Reviewed By: fangyi-zhou

Differential Revision: D82902743

fbshipit-source-id: 34c10c1ccc61b1d62b08eb96805f632a14685fb7