Track another mutable capture bug
Summary:
We should include a test of when `nonlocal` finds a `global` in a parent
scope, in a context with lots of nesting.

This is an edge case, but the runtime behavior is to crash here because
`nonlocal` stops when it finds a name, but then realizes that name is
actually referring to a global.

Reviewed By: fangyi-zhou

Differential Revision: D82902742

fbshipit-source-id: 24cad7134b6bc2b58b5aa5d4a2da1addea2e0c13