Look up mutable captures at static time
Summary:
This is a substantial refactor and is at the heart of the work I'm trying to do
on mutable captures: we now find the original definition in static
initialization instead of after the fact during binding.

It only directly eliminates two unit test bugs - nonsense errors when a `global`
statment happens with an intervening nonlocal name that matches, which is fine
but Pyrefly thought was not legal.

It also fixes two mismatches in conformance tests.

But it also puts plumbing in place to fix other errors as well, and also
eliminates a lot of coupling between `bindings.rs` and scope internals.

Reviewed By: samwgoldman

Differential Revision: D82902736

fbshipit-source-id: eb3bed04f9b242123d373ccdf743c32f7868a04b