Simplify binding-layer mutable capture logic.
Summary:
There's no need for `lookup_mutable_capture` anymore - we can inline that logic
and especially after tweaking things a bit for clarity we get a single,
relatively compact function representing the binding-layer logic for mutable
capture declarations.

I added some comments for good measure, since I find it a little easy to get
mixed up about what the `key` actually represents here (it's confusing because
the `StaticInfo.as_key()` for the mutable capture returns the
`Key::MutableCapture`, which is what we're *defining* here, so it's not that -
it's the key coming from the *original* definition in some enclosing scope.

A short comment goes a long way here.

Reviewed By: samwgoldman

Differential Revision: D82902740

fbshipit-source-id: bbd4b5a6231909e45bb1e95989e2009375af98b2