Add some plumbing to look up mutable captures
Summary:
In order to allow StaticInfo to properly resolve mutable captures, we need
the ability to look up the scope stack to find enclosing scopes when we are
in a nested scope.

Before we get into any actual wiring, it's helpful to modify the plumbing of
`init_current_static` to work the way we will need: the top-level scope should
be modified in place, because scopes is a `Vec1` so we can't pop (and we can
just hardcode all mutable capture lookups to fail, since they are never legal)
whereas for other scopes we can work around the borrow checker by popping the
scope, doing what we need using an `&scopes` for lookups, and pushing it back.

Reviewed By: samwgoldman

Differential Revision: D82902739

fbshipit-source-id: 0c03685feb119ad641ccef6d87a52f508b303944