Re-structure `of_definition` into a flat match
Summary:
With a flat double-match, it's easier to directly represent precedence
that depends on both the style and the count.

As of trunk the double-match is probably only a tiny bit better, but I'm going
to change the MutableCapture vs Anywhere logic shortly in a way that would
make the if/else wrapping a match even messier, so a flat match is much
better going forward.

Reviewed By: samwgoldman

Differential Revision: D82851676

fbshipit-source-id: af1948593803a1e9497239b794f7671a6b9a8c9a