Create `StaticStyle` to express static info more clearly
Summary:
This diff has no actual semantic changes, but completely rewrites the internals
of the existing semantics for `StaticInfo`, by creating a `StaticStyle` which
diverges from `DefinitionStyle` in that it:
- Models multiple definitions differently - in `Definitions` we rely on
  the `count` for this and pick the "minimal" definition, wheras in bindings we
  really just want a marker for when an `Anywhere` is necessary
- Models annotations as `Idx<KeyAnnotation>` instead of bare `ShortIdentifier`
- Combines a number of definition cases that we don't actually care about
  distinguishing.
- Models mutable captures differently, because we will in certain cases pull
  the annotation from a parent scope... the logic for this is not very good
  and is tied to multiple open bugs, but this refactor is a step toward fixing
  all of those.

The correspondance has some amount of compiler pipeline / IR flavor in that the two
variants are related, but they need to be represented in ways that diverge
significantly. I rely exclusively on full pattern matches for the conversion to
avoid any implicit behaviors slipping through unnoticed if we make changes to
`DefinitionStyle` going forward.

This diff is intended to be as minimal as possible, modifying only `scope.rs` itself
to the extent I can and making the absolute minimal changes to other modules.

A few small cleanups are possible, which I'll stack on top for easier review.

Reviewed By: samwgoldman

Differential Revision: D82851680

fbshipit-source-id: 8d6fcccffb5bdc1b70bdd141b6152d0c2f83b7c1