Move `init_current_static` logic down
Summary:
We currently have a chunk of tech debt - both bugs and messy logic
that is in the way of refactors - tied to an inability to correctly
relate mutable captures to the enclosing scope that they come from.

To fix this, we are going to need the `.stmts` method of `Static`
to be able to understand the current scope stack. But that's not possible
under rust ownership rules if we leak a mutable reference to `Scope`
up to the `BindingsBuilder` level, because that mutable reference prevents
us from reading `self.scopes` inside the `stmts` logic.

The most obvious fix I see is to push the core logic down to a layer where
we can work around the borrow checker details.

Reviewed By: samwgoldman

Differential Revision: D82902745

fbshipit-source-id: 9e88dcf1303363c567bb4a05c373ac2ebcb1a2f5