Hide `annot`, use a method
Summary:
I'm about ready to change the layout of `static_info` in ways
that will allow me to simultaneously kill off most remaining
mutable capture scoping bugs and also simplify the API in ways
that will make it easier for me to implement narrowing-and-merging
changes.

One problem is a bit of chicken-and-egg with the "lookup mutable capture"
logic - these functions need to be killed, and will be when I restructure
StaticInfo, but the fact that they directly access `annot` gets in the way
of the very refactors that will allow me to kill them.

Let's just use a method instead.

Reviewed By: samwgoldman

Differential Revision: D82851679

fbshipit-source-id: 64e2dd9aa6a4d970fbfe65f5bdf20c9f13166a2a