Make annotation part of `DefinitionStyle`
Summary:
The current Pyrefly logic doesn't really express invariants around what
kinds of definitions can be annotated.

This is a problem for me - it's true that no direct user of `Definition`
really cares, but putting it into the style makes it clearer which possibilities
are actually reachable, and also will allow us to rely on the same invariants
in `StaticInfo` where we care a lot more about annotations.

It also simplifies the helpers - all but one call of the `add_name.*` functions
was passing `None` as the annotation, now that's no longer neeeded :)

Note that we never drop an annotation, because of the ordering on `DefinitionStyle`

Reviewed By: samwgoldman

Differential Revision: D82851678

fbshipit-source-id: 2beee3c3f23f1ae610404b0f1f6bda1f253b095a