Put all the legacy tparam logic in one place
Summary:
If we move legacy-tparam-relevant methods from `BindingsBuilder` down
to the bottom, we won't have to scroll past irrelevant normal binding
helpers when trying to understand the logic.

It's true that go-to-def doesn't care, but if you jump around a few times
it can be hard to find your way back and scrolling is useful.

Reviewed By: grievejia

Differential Revision: D83020261

fbshipit-source-id: 8b1563757ffb8ea1d6b20099cab7b4eb5aee9ac6