Replace `count` with `needs_anywhere`
Summary:
It was suggested in diff review that we should push the complexity around
deciding when we need `Anywhere` bindings down to the definitions level.

I wasn't comfortable doing this by making `count` do something other than
count, so I prefer switching to a boolean flag that actually represents
the exact thing we care about, which makes it easier to track the behavior
across the bindings / definitions boundary.

There's no behavior change for Pyrefly as a whole here, I tweaked a couple of
unit tests to verify the slightly more subtle behavior at `definitions` time

Reviewed By: grievejia

Differential Revision: D82995751

fbshipit-source-id: ee8e9aa32fb0f8c6c30b4a4787bdb16c0cab1420