Don't push down contextual return hint if incompatible with typevar bound
Summary:
This is related to https://github.com/facebook/pyrefly/issues/881 -- currently return type hints are too aggressive and cause the call's return type to be inferred as the hint, even when the call returns a narrower type.

An especially confusing version of this error is when the hint is incompatible with the typevar's bounds. In this case, the error is very difficult to understand because it's not clear where the type is coming from.

Until we have a better fix to 881, I want to land this to avoid unnecessary confusion.

Reviewed By: stroxler

Differential Revision: D83028875

fbshipit-source-id: 25d1be06c007270d2a7c5add3c924d4b3768e106