Resolve identifier and attribute access callees
Summary:
This is necessary to build correct call graphs for the newly added unit tests.

This diff also creates a macro so that when a test fails, the error message
tells us which test failed. Otherwise there seems no easy way to know which
test fails, which is important since we will have hundred test cases.

Reviewed By: arthaud

Differential Revision: D82998956

fbshipit-source-id: a32fe53cd19aa3339e71816f2fd1e7010f11db29