Make test harness run in opt mode
Summary:
I am noticing the test harness running *significantly* faster with this. Overall, things terminated way faster than before, and this (to me) was a much better experience running the tests.

Additionally, lsp tests timeout occasionally and being in opt mode completely fixes the issue.

I did get a warning:

`mode/opt` was specified, but not found. Using file at `//mode/opt`.

I am not too familiar with this code so I suspect the change isn't 100% correct, but I believe this is running in opt mode, so I can look into getting rid of the warning and the right way to make the change.

Wondering what people's thoughts are and if this is something we could consider?

Reviewed By: ndmitchell

Differential Revision: D82989270

fbshipit-source-id: 9b2b193e8e1dcc8e481b1015cb731eae597f6f97