Add entire django stubs to test harness
Summary:
If we are running in opt mode, (or at least at a higher optimization level), then the whole testsuite takes only 1 min to run with this entire library checked in.

Why not check in the entire stub library? This will ensure consistency against the real library, and we don't have to manually write any stubs for new features moving forward.

Reviewed By: rchen152

Differential Revision: D82856606

fbshipit-source-id: ad180d5af88c8dd8173bc975be416c8f6b86616e