Rearrange class-and-metadata-keys logic
Summary:
It was very weird to me to have a `Scope`-level operation as a static
function in the impl of `Scopes`. Making it a `Scope` method seems
sensible.

But I'd rather not make the method public (one thing I'm trying to do is make
scope modular so that it's much easier to read and refactor because its
internals stay internal and it's public API is declared), and we already have
an `enclosing_class_and_metadata_keys` method so it makes sense to have an
analagously-named method for current scope.

As a result we can also make `current` private, which makes me very happy :)

Reviewed By: grievejia

Differential Revision: D83055325

fbshipit-source-id: 40369e9e4f23a1efd60b1d556a992cd8581f2428