Tweak `ensure_name` functions a bit
Summary:
The `ensure_mutable_name` isn't doing what its name implies anymore, it's just
a bare wrapper around `ensure_name` with a simplified signature for the callsites
where it is used.

I also added a note about why it's even here - naively, it's not adding much to
`ensure_expr`, but the fact that `x` is a regular ref makes this much easier to
use.

And `ensure_name` is no longer in the public API.

Reviewed By: grievejia

Differential Revision: D83069632

fbshipit-source-id: 02ec7750e9058294daf7380f22806e310a1f0164